+++ app/presenters/hotwire_combobox/component/announced.rb
@@ -1,0 +1,5 @@
+module HotwireCombobox::Component::Announced
+  def announcer_attrs
+    { class: "hw-combobox__announcer", data: { hw_combobox_target: "announcer" }, aria: { live: :polite, atomic: true } }
+  end
+end
