+++ app/presenters/hotwire_combobox/component/markup/form.rb
@@ -1,0 +1,6 @@
+module HotwireCombobox::Component::Markup::Form
+  private
+    def form_object
+      form&.object
+    end
+end
