+++ app/presenters/hotwire_combobox/component/markup/handle.rb
@@ -1,0 +1,7 @@
+module HotwireCombobox::Component::Markup::Handle
+  def handle_attrs
+    customize :handle, base: {
+      class: "hw-combobox__handle",
+      data: { action: "click->hw-combobox#clearOrToggleOnHandleClick", hw_combobox_target: "handle" } }
+  end
+end
