+++ app/presenters/hotwire_combobox/component/markup/label.rb
@@ -1,0 +1,5 @@
+module HotwireCombobox::Component::Markup::Label
+  def label_attrs
+    customize :label, base: { class: "hw-combobox__label", for: input_id, hidden: label.blank? }
+  end
+end
