+++ app/presenters/hotwire_combobox/component/markup/wrapper.rb
@@ -1,0 +1,7 @@
+module HotwireCombobox::Component::Markup::Wrapper
+  def main_wrapper_attrs
+    customize :main_wrapper, base: {
+      class: "hw-combobox__main__wrapper",
+      data: { action: ("click->hw-combobox#openByFocusing:self" if multiselect?), hw_combobox_target: "mainWrapper" } }
+  end
+end
