+++ app/presenters/hotwire_combobox/component/multiselect.rb
@@ -1,0 +1,6 @@
+module HotwireCombobox::Component::Multiselect
+  private
+    def multiselect?
+      multiselect_chip_src.present?
+    end
+end
